moved document class to own file
